fachada-fastapi/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
supabase-py==2.3.1
docker==6.1.3
httpx==0.27.0
pydantic==2.6.4
